Search across the keep nodes for timing annotation!
In order to generate a complete set of multi-cycle path constraints, for each
combinational node C, we should generate multi-cycle constraints from
all source registers of the combinational cone rooted on C.

In this process, we will search across the node with 'keep' attribute, i.e.
we will cross the false path boundary. But we do not need to worry about that
the refined timing information protected by the false path boundary will be
lost, because we keep a localized source interval information for each
combinational node that is annotated with timing information.