Since the partial order between live intervals in the CFG is not well defined, remove the non-transitive edge on the compatibility graph to maintain the paritial order.
Specifically, if we found a > b and b > c but a > c is false, we break the
partial order on a > b, because a usually have the bigger live interval
than c, and hence harder to be shared than c. Breaking a > b istead breaking
all c such that b > c but a > c is false can maintain a bigger compatibility
cluster.