Still include the bitmask of a VASTSeqValue into its structural hashing calculation.
Now all VASTSeqValue derived from the same register share the same bitmask,
except the VASTSeqValue derived from the output of memory bus. In this case
the higher part is the alginment of the memory output. Usually we can deduce
these alginments to constant during bit level optimization. And we want to
keep these bitmask in different VASTSeqValue of the same memory bus output.
Hence they are not identical and should not have the same structural hashing,
even though their have the same name.